@@ -1,7 +1,9 @@
 package org.example.relationship.manytoMany.entity;
 
 
-import javax.persistence.*;
+
+import jakarta.persistence.*;
+
 import java.util.Set;
 
 @Entity
