@@ -1,6 +1,9 @@
 package org.example.relationship.manytoMany.entity;
 
-import javax.persistence.*;
+import jakarta.persistence.Entity;
+import jakarta.persistence.*;
+
+
 import java.util.Set;
 
 @Entity
