@@ -1,7 +1,7 @@
 package org.example.relationship.oneToOne.entity;
 
 
-import javax.persistence.*;
+import jakarta.persistence.*;
 
 @Entity
 public class Person {
