@@ -1,7 +1,7 @@
 package org.example.relationship.oneTomany.entity;
 
 
-import javax.persistence.*;
+import jakarta.persistence.*;
 import java.util.HashSet;
 import java.util.Set;
 
