@@ -1,7 +1,6 @@
 package org.example.relationship.oneTomany.entity;
 
-
-import javax.persistence.*;
+import jakarta.persistence.*;
 
 @Entity
 public class PhoneNumber {
